Updated callback methods for logout
