+++ src/server.ts
@@ -189,44 +189,78 @@ export class Server {
   }
 
 // tslint:disable-next-line: no-any
- private logout = async (reqObj: any) => {
-      const keyCloakPropertyName = 'keycloak-token'
-      if (reqObj.session) {
-        logInfo('session exist for reqObj')
-        if (reqObj.session.hasOwnProperty(keyCloakPropertyName)) {
-             const keycloakToken = reqObj.session[keyCloakPropertyName]
-             if (keycloakToken) {
-               const tokenObject = JSON.parse(keycloakToken)
-               const refreshToken = tokenObject.refresh_token
-               if (refreshToken) {
-                 const host = reqObj.get('host')
-                 const urlValue = `https://${host}` + '/auth/realms/' + CONSTANTS.KEYCLOAK_REALM + '/protocol/openid-connect/logout'
-                 try {
-                     request.post({
-                         form: {
-                             client_id: CONSTANTS.KEYCLOAK_GOOGLE_CLIENT_ID,
-                             client_secret: CONSTANTS.KEYCLOAK_GOOGLE_CLIENT_SECRET,
-                             refresh_token: refreshToken,
-                         },
-                         url: urlValue,
-                     })
-                 } catch (err) {
-                     // tslint:disable-next-line: no-console
-                     console.log('Failed to call keycloak logout API ', err, '------', new Date().toString())
-                 }
-               } else {
-                 logError('Not able to retrieve refresh_token value from Session. Logout process failed.')
-               }
-             } else {
-               logError('Not able to retrieve keycloak-token value from Session. Logout process failed.')
-             }
-           } else {
-             logError('Session does not have property with name: ' + keyCloakPropertyName)
-           }
+  private logout = async (reqObj: any) => {
+    const keyCloakPropertyName = 'keycloak-token'
+    if (reqObj.session) {
+      logInfo('session exist for reqObj')
+      if (reqObj.session.hasOwnProperty(keyCloakPropertyName)) {
+        const keycloakToken = reqObj.session[keyCloakPropertyName]
+        if (keycloakToken) {
+          const tokenObject = JSON.parse(keycloakToken)
+          const refreshToken = tokenObject.refresh_token
+          if (refreshToken) {
+            const host = reqObj.get('host')
+            const urlValue = `https://${host}` + '/auth/realms/' + CONSTANTS.KEYCLOAK_REALM + '/protocol/openid-connect/logout'
+            const formData: Record<string, string> = {
+              client_id: 'portal',
+              refresh_token: refreshToken,
+            }
+            if (reqObj.session.hasOwnProperty('keycloakClientId') && (reqObj.session.keycloakClientId !== '')) {
+              formData.client_id = reqObj.session.keycloakClientId
+              formData.client_secret = reqObj.session.keycloakClientSecret
+            }
+            logInfo('formData used in logout: ' + JSON.stringify(formData))
+            try {
+              request.post({
+                form: formData,
+                url: urlValue,
+              })
+            } catch (err) {
+              // tslint:disable-next-line: no-console
+              console.log('Failed to call keycloak logout API ', err, '------', new Date().toString())
+            }
+            if (reqObj.session.parichayToken) {
+              logInfo('Parichay login found... trying to logout from Parichay...')
+              try {
+                request.get({
+                  headers: {
+                    Authorization: reqObj.session.parichayToken.access_token,
+                  },
+                  url: CONSTANTS.PARICHAY_REVOKE_URL,
+                }, (err, res, body) => {
+                  if (err) {
+                    logError('Received error when calling Parichay logout... ')
+                    logError(JSON.stringify(err))
+                  }
+                  if (res) {
+                    logInfo('Received response from Parichay logout... ')
+                    logInfo(JSON.stringify(res.body))
+                  }
+                  if (body) {
+                    logInfo('Received body from Parichay logout...')
+                    logInfo(JSON.stringify(body))
+                  }
+                })
+              } catch (err) {
+                // tslint:disable-next-line: no-console
+                console.log('Failed to call parichay revoke API ', err, '------', new Date().toString())
+              }
+            }
           } else {
-            logError('Session does not exist')
+            logError('Not able to retrieve refresh_token value from Session. Logout process failed.')
           }
-      delete reqObj.session.userRoles
-      delete reqObj.session.userId
+        } else {
+          logError('Not able to retrieve keycloak-token value from Session. Logout process failed.')
+        }
+      } else {
+        logError('Session does not have property with name: ' + keyCloakPropertyName)
+      }
+    } else {
+      logError('Session does not exist')
     }
+    delete reqObj.session.userRoles
+    delete reqObj.session.userId
+    delete reqObj.session.keycloakClientId
+    delete reqObj.session.keycloakClientSecret
+  }
 }
